+++ CHANGELOG.md
@@ -4,7 +4,10 @@
 
 #### 功能构建
 
-- 实现 notify 模块的路由器、调度器、发送器。
+- 实现 `familyhelper-finance` 模块的推送器。
+  - com.dwarfeng.familyhelper.plugin.finance.handler.pusher.FamilyhelperPusher。
+
+- 实现 `notify` 模块的路由器、调度器、发送器。
   - com.dwarfeng.familyhelper.plugin.notify.handler.dispatcher.GeneralDispatcherRegistry。
   - com.dwarfeng.familyhelper.plugin.notify.handler.router.PermissionRouterRegistry。
   - com.dwarfeng.familyhelper.plugin.notify.handler.sender.BuiltinSenderRegistry。